Show url project as link, show username in login screen
